+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - fix(calendar): Add final new line at the end of event description (#162) (!118)
 - fix(meta): Prevent robot indexation (#163) (!120)
 - fix(api): Cleanup OpenAPI security (!122)
+- fix(js): Don't autoreload the app if the network is no available (!123)
 
 ## Changed